добавленно определение местоположения с сохранением. переход на hilt

+++ app/src/main/java/com/example/weatherforecasts/yandexWeather/AnswerYandex.kt
@@ -1,7 +1,7 @@
 package com.example.weatherforecasts.yandexWeather
 
 data class AnswerYandex(
-    val now: Int,
+    val now: Float,
     val now_dt: String,
     val info: Info,
     val fact: Fact,
@@ -15,9 +15,9 @@ data class Info(
 )
 
 data class Fact(
-    val temp: Int,
-    val feels_like: Int,
-    val temp_water: Int,
+    val temp: Float,
+    val feels_like: Float,
+    val temp_water: Float,
     val icon: String,
     val condition: String,
     val wind_speed: Float,
@@ -25,29 +25,29 @@ data class Fact(
     val wind_dir: String,
     val pressure_mm: Float,
     val pressure_pa: Float,
-    val humidity: Int,
+    val humidity: Float,
     val daytime: String,
     val polar: Boolean,
     val season: String,
-    val obs_time: Int
+    val obs_time: Float
 )
 
 data class Forecast(
     val date: String,
-    val date_ts: Int,
-    val week: Int,
+    val date_ts: Float,
+    val week: Float,
     val sunset: String,
-    val moon_code: Int,
+    val moon_code: Float,
     val moon_text: String,
     val parts: List<Parts>
 )
 
 data class Parts(
     val part_name: String,
-    val temp_min: Int,
-    val temp_max: Int,
-    val temp_avg: Int,
-    val feels_like: Int,
+    val temp_min: Float,
+    val temp_max: Float,
+    val temp_avg: Float,
+    val feels_like: Float,
     val icon: String,
     val condition: String,
     val daytime: String,
@@ -57,8 +57,8 @@ data class Parts(
     val wind_dir: String,
     val pressure_mm: Float,
     val pressure_pa: Float,
-    val humidity: Int,
+    val humidity: Float,
     val prec_mm: Float,
-    val prec_period: Int,
+    val prec_period: Float,
     val prec_prob: Float
 )
